Image.getImageWidth/Height() should return int, not double
This isn’t strictly a non-breaking change: it might select a different
method overload or inferred var type in contexts that call it. However,
it should be non-breaking _enough_ for real-world Kilt Graphics usage.

+++ src/edu/macalester/graphics/Image.java
@@ -202,16 +202,16 @@ public class Image extends GraphicsObject {
     }
 
     /**
-     * Get the width of the underlying image.
+     * Get the width of the underlying image in pixels.
      */
-    public double getImageWidth() {
+    public int getImageWidth() {
         return img == null ? 0 : img.getWidth();
     }
 
     /**
-     * Get the height of the underlying image.
+     * Get the height of the underlying image in pixels.
      */
-    public double getImageHeight() {
+    public int getImageHeight() {
         return img == null ? 0 : img.getHeight();
     }
